reenable unit test scaffolding; run unit tests as part of CI
