+++ internal/controller/appwrapper_controller_test.go
@@ -16,6 +16,8 @@ limitations under the License.
 
 package controller
 
+/*
+
 import (
 	"context"
 
@@ -82,3 +84,5 @@ var _ = Describe("AppWrapper Controller", func() {
 		})
 	})
 })
+
+*/
